requirements.txt
requests
google-generativeai
python-dotenv
pytest
